+++ R/mlr_reflections.R
@@ -146,7 +146,7 @@ local({
   mlr_reflections$resampling_properties = c("duplicated_ids", "weights")
 
   ### Measures
-  tmp = c("na_score", "requires_task", "requires_learner", "requires_model", "requires_train_set", "weights")
+  tmp = c("na_score", "requires_task", "requires_learner", "requires_model", "requires_train_set", "weights", "primary_iters")
   mlr_reflections$measure_properties = list(
     classif = tmp,
     regr = tmp